+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # All changes will be documented here.
 
+- Add support for deploying cloudwatch agent
 - Support basic configuration for an EC2 instance.
 - Adding support for ssm configuration.